Remove set block interval interface (#900)
* update interaction submodule

* Revert "Merge pull request #857 from ouwenkg/upgrade-scripts"

This reverts commit 29897d29da1ed980f07d04dcae9516fdd4bd749d, reversing
changes made to 24f1649114a6071c61f3febfd23cab0d61bac2d4.

* Revert "Merge pull request #855 from
ouwenkg/interface-block-interval"[skip audit]

This reverts commit 24f1649114a6071c61f3febfd23cab0d61bac2d4, reversing
changes made to 9d209c2cb8a6cccfe7aa0e82708a859507bb9e29.
