+++ scripts/amend_system_contracts.py
@@ -1,13 +1,25 @@
+#!/usr/bin/env python3
+# -*- coding:utf-8 -*-
+"""
+Amend the code and storage of system contract.
+"""
 
+import argparse
+import functools
+import json
+import logging
+import subprocess
 import sys
 import time
-import logging
-import functools
-import subprocess
 from jsonrpcclient.http_client import HTTPClient
 
-LOCAL = 'http://127.0.0.1:1337'
+LATEST_VERSION = 2
 AMEND_ADDR = '0xffffffffffffffffffffffffffffffffff010002'
+SYS_CONF = '0xffffffffffffffffffffffffffffffffff020000'
+NEW = '0xffffffffffffffffffffffffffffffffff020012'
+KEY = '0x30'
+LOCAL = 'http://127.0.0.1:1337'
+
 
 def send_tx(args):
     """
@@ -26,6 +38,8 @@ def send_tx(args):
         '--to': AMEND_ADDR,
         '--code': args.code,
         '--value': str(args.value),
+        '--version': str(args.version),
+        '--chain_id': str(args.chain_id),
     }
     args = functools.reduce(
         lambda lst, kv: lst + list(kv),
@@ -38,10 +52,12 @@ def send_tx(args):
     with open('../output/transaction/hash') as fobj:
         return fobj.read().strip()
 
+
 def rpc_request(method, params, url=LOCAL):
     """ Send a jsonrpc request to default url. """
     client = HTTPClient(url)
     return client.request(method, params)
+
 
 def get_receipt(tx_hash, url, retry=8):
     """ Get receipt of a transaction """
@@ -51,6 +67,7 @@ def get_receipt(tx_hash, url, retry=8):
             return receipt
         time.sleep(4)
         retry -= 1
+
 
 def amend_code(addr, code, args):
     """ Amend the code """
@@ -69,6 +86,7 @@ def amend_code(addr, code, args):
         logging.critical('amend code of %s exception: %s', addr, exception)
         sys.exit(1)
 
+
 def amend_storage(addr, key, val, args):
     """ Amend the storage: key and value """
     try:
@@ -85,3 +103,49 @@ def amend_storage(addr, key, val, args):
         logging.critical('amend storage of %s[%s] exception: %s', addr, key,
                          exception)
         sys.exit(1)
+
+
+def parse_arguments():
+    """ parse the arguments: chain_id, version, privkey, url """
+    parser = argparse.ArgumentParser()
+    parser.add_argument(
+        "--version",
+        help="Tansaction version.",
+        default=LATEST_VERSION,
+        type=int)
+    parser.add_argument(
+        '--privkey', required=True, help='The admin private key.')
+    parser.add_argument(
+        '--chain_id', required=True, help='The id of the chain.')
+    parser.add_argument('--url', required=True, help='The url of the chain.')
+
+    args = parser.parse_args()
+    return args
+
+
+def main():
+    """ Load the genesis file and amend """
+    args = parse_arguments()
+
+    with open('../../genesis.json', 'r') as gene:
+        genesis = json.load(gene)
+
+    alloc = genesis['alloc']
+
+    for addr in alloc:
+        # amend storage
+        args.value = 3
+        storage = alloc[addr]['storage']
+        for key in storage:
+            if addr == NEW:
+                amend_storage(addr, key, storage[key], args)
+            elif addr == SYS_CONF and key == KEY:
+                amend_storage(addr, key, storage[key], args)
+
+        # amend code
+        args.value = 2
+        amend_code(addr, alloc[addr]['code'], args)
+
+
+if __name__ == '__main__':
+    main()
