+++ tests/integrate_test/cita_features_test.sh
@@ -23,13 +23,6 @@ test_perm_denied() {
     cd ../../..
 }
 
-test_change_block_interval() {
-    local version=$1
-    cd ./scripts/txtool/txtool
-    python3 "${SOURCE_DIR}"/tests/integrate_test/test_block_interval.py --version="$version"
-    cd ../../..
-}
-
 main() {
     echo -n "0) prepare  ...  "
     # shellcheck source=/dev/null
@@ -61,13 +54,6 @@ main() {
 
     echo -n "5) Check permission denied ... "
     test_perm_denied 2
-    echo "Done"
-
-    echo -e "6) Run block interval tests ...\n"
-    test_change_block_interval 2
-    timeout=$(check_height_growth_normal 0 60) && (echo "FAILED"
-                                                    echo "failed to change block interval"
-                                                    exit 1)
     echo "Done"
 }
